% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{bins_100}
\alias{bins_100}
\title{Exonic Features:GencodeV25:100bp}
\format{A \code{GRanges} object of sub-exonic features.}
\description{
This is \code{GRanges} object of length 1,024,775. The segments 
reflect the sub-exonic pieces calculated by recountNNLS to be used for 
calculating the sufficient statistics of linear modeling.
It is suitable to mode experiments with read lengths around 100bp.
}
\examples{
\dontrun{
data(bins_100)
bins_100
}
}
\author{
Jack Fu  \email{jmfu@jhsph.edu}
}
