% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{bins_150}
\alias{bins_150}
\title{Exonic Features:GencodeV25:150bp}
\format{A \code{GRanges} object of sub-exonic features.}
\description{
This is \code{GRanges} object of length 793,834. The segments 
reflect the sub-exonic pieces calculated by recountNNLS to be used for 
calculating the sufficient statistics of linear modeling.
It is suitable to mode experiments with read lengths around 150bp.
}
\examples{
\dontrun{
data(bins_150)
bins_150
}
}
\author{
Jack Fu  \email{jmfu@jhsph.edu}
}
