% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{bins_37}
\alias{bins_37}
\title{Exonic Features:GencodeV25:37bp}
\format{A \code{GRanges} object of sub-exonic features.}
\description{
This is \code{GRanges} object of length 2,529,474. The segments 
reflect the sub-exonic pieces calculated by recountNNLS to be used for 
calculating the sufficient statistics of linear modeling.
It is suitable to mode experiments with read lengths around 37bp.
}
\examples{
\dontrun{
data(bins_37)
bins_37
}
}
\author{
Jack Fu  \email{jmfu@jhsph.edu}
}
