% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{bins_50}
\alias{bins_50}
\title{Exonic Features:GencodeV25:50bp}
\format{A \code{GRanges} object of sub-exonic features.}
\description{
This is \code{GRanges} object of length 1,877,008. The segments 
reflect the sub-exonic pieces calculated by recountNNLS to be used for 
calculating the sufficient statistics of linear modeling.
It is suitable to mode experiments with read lengths around 50bp.
}
\examples{
\dontrun{
data(bins_50)
bins_50
}
}
\author{
Jack Fu  \email{jmfu@jhsph.edu}
}
