% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{bins_75}
\alias{bins_75}
\title{Exonic Features:GencodeV25:75bp}
\format{A \code{GRanges} object of sub-exonic features.}
\description{
This is \code{GRanges} object of length 1,289,851. The segments 
reflect the sub-exonic pieces calculated by recountNNLS to be used for 
calculating the sufficient statistics of linear modeling.
It is suitable to mode experiments with read lengths around 75bp.
}
\examples{
\dontrun{
data(bins_75)
bins_75
}
}
\author{
Jack Fu  \email{jmfu@jhsph.edu}
}
