% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{gff_jx}
\alias{gff_jx}
\title{Splice Junctions:GencodeV25}
\format{A \code{GRanges} object.}
\description{
This is a \code{GRanges} object containing 293,662 entries. 
Each entry corresponds to an annotated junction based on the GnecodeV25 
reference transcriptome. Strand-specific.
}
\examples{
\dontrun{
data(gff_jx)
gff_jx
}
}
\author{
Jack Fu  \email{jmfu@jhsph.edu}
}
