% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{matrix_75}
\alias{matrix_75}
\title{Emission probability matrices:GencodeV25:75bp}
\format{A \code{list} of length 19,932, where each entry is a matrix of 
variable dimension corresponding to the name. Each row of the matrix 
corresponds to a feature, and each column corresponds to a transcript.}
\description{
This is a \code{list} of the 19,932 emission probability 
matrices that correspond to the coding genes of the GencodeV25 transcriptome 
annotation. It is used to fit recountNNLS regression model for transcript 
abundance estimation of experiments with 75bp read lengths.
}
\examples{
\dontrun{
data(matrix_75)
head(names(matrix_75))
head(matrix_75[[1]])
}
}
\author{
Jack Fu  \email{jmfu@jhsph.edu}
}
