% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{tcga_meta}
\alias{tcga_meta}
\title{TCGA Metadata}
\description{
This is a \code{data.frame} containing the meta information for
TCGA samples that mirror the format output by \code{processPheno} for all
other SRA projects.
}
\examples{
\dontrun{
data(tcga_meta)
dim(tcga_meta)
}
}
\author{
Jack Fu \email{jmfu@jhsph.edu}
}
