% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{tx_grl}
\alias{tx_grl}
\title{Transcript List:GencodeV25}
\format{A \code{GRanges} object.}
\description{
This is a \code{GRangesList} object of length 145,455. Each 
entry of the list conveys the full exon structure of the transcript, 
based on the GencdoeV25 reference transcriptome.
}
\examples{
\dontrun{
data(gff_jx)
gff_jx
}
}
\author{
Jack Fu  \email{jmfu@jhsph.edu}
}
