% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{tx_info}
\alias{tx_info}
\title{Transcript Metadata:GencodeV25}
\description{
This is a \code{data.frame} containing the meta information for
transcripts and genes of the protein coding trasncripts part of the 
GencodeV25 annotation. It is to be packaged in rowData of the
output rse from recountNNLS.
}
\examples{
\dontrun{
data(tx_info)
dim(tx_info)
}
}
\author{
Jack Fu \email{jmfu@jhsph.edu}
\describe{
 \item{\code{tx_id}}{Transcript number.}
 \item{\code{tx_name}}{Transcript identifier.}
 \item{\code{gene_id}}{Gene identifier.}
 \item{\code{nexon}}{Number of exons in transcript.}
 \item{\code{tx_len}}{Length of transcript.}
 }
}
